python/requirements.txt
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
xgboost>=2.0
grpcio>=1.60
grpcio-tools>=1.60
protobuf>=4.25
joblib>=1.3
requests>=2.31
matplotlib>=3.7
